ai_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pinecone-client==4.1.2
google-generativeai==0.5.0
python-dotenv==1.0.0
pydantic==2.5.0
